refactor: Déplacer la logique de gestion des bibs vers une action serveur dédiée pour une meilleure organisation du code
